Work on adding functionality to 'show my hangs' button
We grabbed the ID of the current signed in user. We are trying to
iterate through each hang to compare the ID of the current
signed in user to each hang's owner. If they match, we will have
handlebars show the matching hangs (aka, hangs that the user created).

Co-authored by: Trey Hilson, Brandon DePalmaFarr, Ross Northrup